Move WindowTooSmall to views
