+++ frontend/src/views/WindowTooSmall.js
